Projects/matric_result_checker_agent/pyproject.toml
[project]
name = "matric-result-checker-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.3.3",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
